backend/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-cors-headers==4.3.1
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
setuptools==69.0.3
simplejson==3.19.2
sqlparse==0.4.4
twilio==8.13.0
uritemplate==4.1.1
urllib3==2.2.0
yarl==1.9.4
